requirements.txt
PyQt5>=5.15,<5.16
paho-mqtt>=1.6